# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: DevSync REST API
  description: Self-hosted local developer workspace and high-speed file sharing APIs.
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /files:
    get:
      summary: List and search shared files
      tags:
        - File Sharing
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of files
    delete:
      summary: Delete a file
      tags:
        - File Sharing
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File deleted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT